move action plan presenter code into a shared presenter class
