+++ server/routes/serviceProviderReferrals/interventionProgressPresenter.ts
@@ -6,6 +6,7 @@ import DateUtils from '../../utils/dateUtils'
 import sessionStatus, { SessionStatus } from '../../utils/sessionStatus'
 import SessionStatusPresenter from '../shared/sessionStatusPresenter'
 import Intervention from '../../models/intervention'
+import ActionPlanDetailsPresenter from '../shared/actionPlanDetailsPresenter'
 
 interface EndedFields {
   endRequestedAt: string | null
@@ -23,57 +24,30 @@ interface ProgressSessionTableRow {
 export default class InterventionProgressPresenter {
   referralOverviewPagePresenter: ReferralOverviewPagePresenter
 
+  actionPlanDetailsPresenter: ActionPlanDetailsPresenter
+
   constructor(
     private readonly referral: SentReferral,
     private readonly intervention: Intervention,
     private readonly actionPlan: ActionPlan | null,
     private readonly actionPlanAppointments: ActionPlanAppointment[]
   ) {
+    const subNavUrlPrefix = 'service-provider'
     this.referralOverviewPagePresenter = new ReferralOverviewPagePresenter(
       ReferralOverviewPageSection.Progress,
       referral.id,
-      'service-provider'
+      subNavUrlPrefix
     )
+    this.actionPlanDetailsPresenter = new ActionPlanDetailsPresenter(actionPlan, subNavUrlPrefix)
   }
 
   get referralAssigned(): boolean {
     return this.referral.assignedTo !== null
   }
 
-  readonly actionPlanFormUrl = `/service-provider/action-plan/${this.actionPlan?.id}/add-activities`
-
-  readonly createActionPlanFormAction = `/service-provider/referrals/${this.referral.id}/action-plan`
-
   readonly text = {
     title: utils.convertToTitleCase(this.intervention.contractType.name),
-    actionPlanStatus: this.actionPlanStatus,
-    actionPlanSubmittedDate: DateUtils.getDateStringFromDateTimeString(this.actionPlan?.submittedAt || null),
-    // fixme: we don't have immediate access to the full name of the user without looking it up ahead of time
-    // actionPlanSubmitter: this.actionPlan?.submittedBy.username\
-    actionPlanApprovalDate: DateUtils.getDateStringFromDateTimeString(this.actionPlan?.approvedAt || null),
     endOfServiceReportStatus: this.endOfServiceReportSubmitted ? 'Submitted' : 'Not submitted',
-  }
-
-  private get actionPlanStatus(): string {
-    if (this.actionPlanApproved) {
-      return 'Approved'
-    }
-    if (this.actionPlanUnderReview) {
-      return 'Under review'
-    }
-    return 'Not submitted'
-  }
-
-  get actionPlanCreated(): boolean {
-    return this.actionPlan !== null
-  }
-
-  get actionPlanUnderReview(): boolean {
-    return this.actionPlan?.submittedAt != null && !this.actionPlanApproved
-  }
-
-  get actionPlanApproved(): boolean {
-    return this.actionPlan?.approvedAt != null
   }
 
   get referralEnded(): boolean {
