+++ server/routes/serviceProviderReferrals/interventionProgressView.ts
@@ -80,7 +80,7 @@ export default class InterventionProgressView {
         key: { text: 'Action plan status' },
         value: {
           text: tagMacro({
-            text: this.presenter.text.actionPlanStatus,
+            text: this.presenter.actionPlanDetailsPresenter.text.actionPlanStatus,
             classes: this.actionPlanTagClass,
             attributes: { id: 'action-plan-status' },
           }),
@@ -88,12 +88,14 @@ export default class InterventionProgressView {
       },
     ]
 
-    if (!this.presenter.actionPlanCreated) {
+    if (!this.presenter.actionPlanDetailsPresenter.actionPlanCreated) {
       // action plan doesn't exist; show link to create one
       rows.push({
         key: { text: 'To do' },
         value: {
-          html: `<form method="post" action="${ViewUtils.escape(this.presenter.createActionPlanFormAction)}">
+          html: `<form method="post" action="${ViewUtils.escape(
+            this.presenter.actionPlanDetailsPresenter.createActionPlanFormAction
+          )}">
                    <input type="hidden" name="_csrf" value="${ViewUtils.escape(csrfToken)}">
                    <button class="govuk-button govuk-button--secondary">
                      Create action plan
@@ -101,37 +103,51 @@ export default class InterventionProgressView {
                  </form>`,
         },
       })
-    } else if (this.presenter.actionPlanCreated && !this.presenter.actionPlanUnderReview) {
+    } else if (
+      this.presenter.actionPlanDetailsPresenter.actionPlanCreated &&
+      !this.presenter.actionPlanDetailsPresenter.actionPlanUnderReview
+    ) {
       // action plan exists, but has not been submitted
       rows.push({
         key: { text: 'To do' },
-        value: { html: `<a href="${this.presenter.actionPlanFormUrl}" class="govuk-link">Submit action plan</a>` },
+        value: {
+          html: `<a href="${this.presenter.actionPlanDetailsPresenter.actionPlanFormUrl}" class="govuk-link">Submit action plan</a>`,
+        },
       })
-    } else if (this.presenter.actionPlanUnderReview) {
+    } else if (this.presenter.actionPlanDetailsPresenter.actionPlanUnderReview) {
       // action plan has been submitted; show link to view it
       rows.push({
         key: { text: 'Submitted date' },
-        value: { text: this.presenter.text.actionPlanSubmittedDate, classes: 'action-plan-submitted-date' },
+        value: {
+          text: this.presenter.actionPlanDetailsPresenter.text.actionPlanSubmittedDate,
+          classes: 'action-plan-submitted-date',
+        },
       })
       // FIXME: the 'view action plan' page doesn't exist yet!
       rows.push({
         key: { text: 'To do' },
-        value: { html: `<a href="#" class="govuk-link">View action plan</a>` },
+        value: {
+          html: `<a href="${this.presenter.actionPlanDetailsPresenter.viewActionPlanUrl}" class="govuk-link">View action plan</a>`,
+        },
       })
-    } else if (this.presenter.actionPlanApproved) {
+    } else if (this.presenter.actionPlanDetailsPresenter.actionPlanApproved) {
       // action plan has been approved; show link to view it
       rows.push({
         key: { text: 'Submitted date' },
-        value: { text: this.presenter.text.actionPlanSubmittedDate, classes: 'action-plan-submitted-date' },
+        value: {
+          text: this.presenter.actionPlanDetailsPresenter.text.actionPlanSubmittedDate,
+          classes: 'action-plan-submitted-date',
+        },
       })
       rows.push({
         key: { text: 'Approval date' },
-        value: { text: this.presenter.text.actionPlanApprovalDate },
+        value: { text: this.presenter.actionPlanDetailsPresenter.text.actionPlanApprovalDate },
       })
-      // FIXME: the 'view action plan' page doesn't exist yet!
       rows.push({
         key: { text: 'To do' },
-        value: { html: `<a href="#" class="govuk-link">View action plan</a>` },
+        value: {
+          html: `<a href="${this.presenter.actionPlanDetailsPresenter.viewActionPlanUrl}" class="govuk-link">View action plan</a>`,
+        },
       })
     }
 
@@ -139,7 +155,7 @@ export default class InterventionProgressView {
   }
 
   private get actionPlanTagClass(): string {
-    switch (this.presenter.text.actionPlanStatus) {
+    switch (this.presenter.actionPlanDetailsPresenter.text.actionPlanStatus) {
       case 'Approved':
         return 'govuk-tag--green'
       case 'Under review':
